+++ src/main/java/ac_one/gqw1024/community/ac_one_community/dto/QuestionDto.java
@@ -9,8 +9,8 @@ import lombok.Data;
  */
 @Data
 public class QuestionDto {
-    private Integer id;//问题的id
-    private Integer creator;//问题创建者，提出该问题的用户的account_id
+    private Long id;//问题的id
+    private Long creator;//问题创建者，提出该问题的用户的account_id
     private String title;//问题的标题
     private String questionType = "default_type";//问题的类型
     private String description;//问题的内容
